Move value(), check_if_numeric_type() to common.numeric_types
