+++ pyomo/core/expr/numvalue.py
@@ -68,6 +68,8 @@ from pyomo.common.numeric_types import (
     native_integer_types,
     native_logical_types,
     pyomo_constant_types,
+    check_if_numeric_type,
+    value,
 )
 from pyomo.core.pyomoobject import PyomoObject
 from pyomo.core.expr.expr_errors import TemplateExpressionError
@@ -145,93 +147,6 @@ class NonNumericValue(object):
 
 
 nonpyomo_leaf_types.add(NonNumericValue)
-
-
-def value(obj, exception=True):
-    """
-    A utility function that returns the value of a Pyomo object or
-    expression.
-
-    Args:
-        obj: The argument to evaluate. If it is None, a
-            string, or any other primitive numeric type,
-            then this function simply returns the argument.
-            Otherwise, if the argument is a NumericValue
-            then the __call__ method is executed.
-        exception (bool): If :const:`True`, then an exception should
-            be raised when instances of NumericValue fail to
-            evaluate due to one or more objects not being
-            initialized to a numeric value (e.g, one or more
-            variables in an algebraic expression having the
-            value None). If :const:`False`, then the function
-            returns :const:`None` when an exception occurs.
-            Default is True.
-
-    Returns: A numeric value or None.
-    """
-    if obj.__class__ in native_types:
-        return obj
-    if obj.__class__ in pyomo_constant_types:
-        #
-        # I'm commenting this out for now, but I think we should never expect
-        # to see a numeric constant with value None.
-        #
-        # if exception and obj.value is None:
-        #    raise ValueError(
-        #        "No value for uninitialized NumericConstant object %s"
-        #        % (obj.name,))
-        return obj.value
-    #
-    # Test if we have a duck typed Pyomo expression
-    #
-    try:
-        obj.is_numeric_type()
-    except AttributeError:
-        #
-        # TODO: Historically we checked for new *numeric* types and
-        # raised exceptions for anything else.  That is inconsistent
-        # with allowing native_types like None/str/bool to be returned
-        # from value().  We should revisit if that is worthwhile to do
-        # here.
-        #
-        if check_if_numeric_type(obj):
-            return obj
-        else:
-            if not exception:
-                return None
-            raise TypeError(
-                "Cannot evaluate object with unknown type: %s" % obj.__class__.__name__
-            ) from None
-    #
-    # Evaluate the expression object
-    #
-    if exception:
-        #
-        # Here, we try to catch the exception
-        #
-        try:
-            tmp = obj(exception=True)
-            if tmp is None:
-                raise ValueError(
-                    "No value for uninitialized NumericValue object %s" % (obj.name,)
-                )
-            return tmp
-        except TemplateExpressionError:
-            # Template expressions work by catching this error type. So
-            # we should defer this error handling and not log an error
-            # message.
-            raise
-        except:
-            logger.error(
-                "evaluating object as numeric value: %s\n    (object: %s)\n%s"
-                % (obj, type(obj), sys.exc_info()[1])
-            )
-            raise
-    else:
-        #
-        # Here, we do not try to catch the exception
-        #
-        return obj(exception=False)
 
 
 def is_constant(obj):
@@ -488,54 +403,6 @@ def as_numeric(obj):
         "Cannot treat the value '%s' as a numeric value because it has "
         "unknown type '%s'" % (str(obj), type(obj).__name__)
     )
-
-
-def check_if_numeric_type(obj):
-    """Test if the argument behaves like a numeric type.
-
-    We check for "numeric types" by checking if we can add zero to it
-    without changing the object's type.  If that works, then we register
-    the type in native_numeric_types.
-
-    """
-    obj_class = obj.__class__
-    # Do not re-evaluate known native types
-    if obj_class in native_types:
-        return obj_class in native_numeric_types
-
-    try:
-        obj_plus_0 = obj + 0
-        obj_p0_class = obj_plus_0.__class__
-        # ensure that the object is comparable to 0 in a meaningful way
-        # (among other things, this prevents numpy.ndarray objects from
-        # being added to native_numeric_types)
-        if not ((obj < 0) ^ (obj >= 0)):
-            return False
-        # Native types *must* be hashable
-        hash(obj)
-    except:
-        return False
-    if obj_p0_class is obj_class or obj_p0_class in native_numeric_types:
-        #
-        # If we get here, this is a reasonably well-behaving
-        # numeric type: add it to the native numeric types
-        # so that future lookups will be faster.
-        #
-        _numeric_types.RegisterNumericType(obj_class)
-        #
-        # Generate a warning, since Pyomo's management of third-party
-        # numeric types is more robust when registering explicitly.
-        #
-        logger.warning(
-            f"""Dynamically registering the following numeric type:
-    {obj_class.__module__}.{obj_class.__name__}
-Dynamic registration is supported for convenience, but there are known
-limitations to this approach.  We recommend explicitly registering
-numeric types using RegisterNumericType() or RegisterIntegerType()."""
-        )
-        return True
-    else:
-        return False
 
 
 @deprecated(
