+++ pyomo/network/port.py
@@ -20,6 +20,7 @@ from pyomo.common.deprecation import RenamedClass
 from pyomo.common.formatting import tabular_writer
 from pyomo.common.log import is_debug_set
 from pyomo.common.modeling import unique_component_name, NOTSET
+from pyomo.common.numeric_types import value
 from pyomo.common.timing import ConstructionTimer
 
 from pyomo.core.base.var import Var
@@ -28,7 +29,7 @@ from pyomo.core.base.component import ComponentData, ModelComponentFactory
 from pyomo.core.base.global_set import UnindexedComponent_index
 from pyomo.core.base.indexed_component import IndexedComponent, UnindexedComponent_set
 from pyomo.core.base.misc import apply_indexed_rule
-from pyomo.core.base.numvalue import as_numeric, value
+from pyomo.core.expr.numvalue import as_numeric
 from pyomo.core.expr.current import identify_variables
 from pyomo.core.base.label import alphanum_label_from_name
