+++ pyomo/repn/standard_repn.py
@@ -18,15 +18,16 @@ import sys
 import logging
 import itertools
 
+from pyomo.common.numeric_types import native_numeric_types
 from pyomo.core.base import Constraint, Objective, ComponentMap
 
 from pyomo.core.expr import current as EXPR
+from pyomo.core.expr.numvalue import NumericConstant
 from pyomo.core.base.objective import _GeneralObjectiveData, ScalarObjective
 from pyomo.core.base import _ExpressionData, Expression
 from pyomo.core.base.expression import ScalarExpression, _GeneralExpressionData
 from pyomo.core.base.var import ScalarVar, Var, _GeneralVarData, value
 from pyomo.core.base.param import ScalarParam, _ParamData
-from pyomo.core.base.numvalue import NumericConstant, native_numeric_types
 from pyomo.core.kernel.expression import expression, noclone
 from pyomo.core.kernel.variable import IVariable, variable
 from pyomo.core.kernel.objective import objective
